% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{auto_production}
\alias{auto_production}
\title{Auto production}
\usage{
auto_production(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns auto production in Mexico and year-over-year change.
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token <- "webservice_token"
Autos <- auto_production(token)
}
}
\author{
Eduardo Flores
}

