% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Autos}
\alias{Autos}
\title{Obtener Produccion de Autos}
\usage{
Autos(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene producción automotriz en México y cambio porcentual anual.
Es un wrapper de las funciones \code{Serie_Inegi()} y \code{YoY()}.
}
\note{
Encoding no permite acentos en titulo de descripción
}
\examples{
ProduccionAutos<-Autos(token)
}
\author{
Eduardo Flores
}

