% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Balanza_Pagos}
\alias{Balanza_Pagos}
\title{Obtener Balanza de Pagos}
\usage{
Balanza_Pagos(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene principales componentes de la Balanza de Pagos: 2 de la Cuenta Corriente, 3 de la Cuenta Financiera y sus 2 resultados.
Es un wrapper de las funciones \code{Serie_Inegi()} y \code{YoY()}.
}
\examples{
BalanzadePagosMexico<-Balanza_Pagos(token)
}
\author{
Eduardo Flores
}

