% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{BoP}
\alias{BoP}
\title{Obtener Balanza de Pagos}
\usage{
BoP(token)

BoP(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}

\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame con 8 columnas

Data.frame
}
\description{
Obtiene principales componentes de la Balanza de Pagos: 2 de la Cuenta Corriente, 3 de la Cuenta Financiera y sus 2 resultados.

Obtiene principales componentes de la Balanza de Pagos: 2 de la Cuenta Corriente, 3 de la Cuenta Financiera y sus 2 resultados.
}
\examples{
BalanzadePagosMexico<-BoP(token)
BalanzadePagosMexico<-BoP(token)
}
\author{
Eduardo Flores

Eduardo Flores
}

