% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Catalogo_Inegi.R
\docType{data}
\name{Catalogo_Inegi}
\alias{Catalogo_Inegi}
\title{Catalogo de algunos indicadores del INEGI}
\format{data.frame with 53940 rows and 10 variables:
\describe{
  \item{ShortName}{Nombre corto de indicador, en minusculas siempre}
  \item{Entidad}{Nacional o Estado}
  \item{Desc}{Descripción oficial del INEGI}
  \item{Link}{Liga directa (url)}
}}
\source{
\url{http://www.inegi.org.mx/desarrolladores/indicadores/apiindicadores.aspx}
}
\usage{
Catalogo_Inegi
}
\description{
Relación de Indicadores y ligas del API del INEGI
}
\keyword{datasets}

