% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_city_inflation.R
\name{city_inflation}
\alias{city_inflation}
\alias{inflacion_ciudades}
\title{Inflation by city}
\usage{
city_inflation(token)

inflacion_ciudades(token)
}
\arguments{
\item{token}{token API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns monthly year-over-year inflation rates for 46 main cities in Mexico.
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
cities <- city_inflation(token)
}
}
\author{
Eduardo Flores
}

