% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_aux.R
\name{compact_inegi_series}
\alias{compact_inegi_series}
\title{Compacts metadata into a data.frame}
\usage{
compact_inegi_series(series, token)
}
\arguments{
\item{series}{INEGI series as passed to \code{inegi_series()}}

\item{token}{INEGI API token}
}
\description{
Returns data.frame with metadata and data from \code{inegi_series()} in data.frame form. Each metadata data is replicated in its corresponding column.
}
\examples{
\dontrun{
df <- compact_inegi_series(GDP_seriescode, token)
}

}
\author{
Eduardo Flores
}
