% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_country_exports.R
\name{country_exports}
\alias{country_exports}
\alias{exports_country}
\alias{series_exportaciones_pais}
\title{Exports by country}
\usage{
exports_country(token)

series_exportaciones_pais(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns exports to main trading partners of all products. Regions are the following: United States, Canada, China, CentralAmerica, SouthAmerica
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token <- "webservice_token"
xbycountry <- exports_country(token)
}
}
\author{
Eduardo Flores
}
