% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_aux.R
\name{crecer}
\alias{crecer}
\title{Grows a series by a set rate}
\usage{
crecer(tasas, comienzo)
}
\arguments{
\item{tasas}{vector with rates}

\item{comienzo}{initial value}
}
\value{
numeric
}
\description{
When specifying an initial starting value, this "grows" the value by a vector of growth rates. This is a legacy function.
}
\examples{
rates <- c(1.10,1.20,1.05,1.02,1.10)

# Grow by that rate
Results <- crecer(tasas = rates, comienzo = 100)

}
\seealso{
series_crecimiento_regiones
}
\author{
Eduardo Flores
}
