% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_econ_sectors.R
\name{econ_sectors}
\alias{econ_sectors}
\alias{rate_econsectors}
\alias{tasa_sectoresYoY}
\title{Sectorial growth rates}
\usage{
rate_econsectors(token)

tasa_sectoresYoY(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns growth rate by economic sector as defined in INEGI (subsectors of IGAE). 
None of the series are seasonally adjusted and percent change is year over year.
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
sectors <- rate_econsectors(token)
}
}
\author{
Eduardo Flores
}

