% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_exchange_rate.R
\name{exchange_rate}
\alias{exchange_rate}
\alias{series_tipocambio}
\title{Exchange rate}
\usage{
exchange_rate(token)

series_tipocambio(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns exchange rate (interbank, sale) for pesos to U.S. dollars.
}
\examples{
\dontrun{
token <- "webservice_token"
USD <- series_tipocambio(token)
USD <- exchange_rate(token)
}
}
\author{
Eduardo Flores
}

