% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{exports_country}
\alias{exports_country}
\title{Exports by country}
\usage{
exports_country(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns exports to main trading partners of all products. Regions are the following: United States, Canada, China, CentralAmerica, SouthAmerica
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token <- "webservice_token"
xbycountry <- exports_country(token)
}
}
\author{
Eduardo Flores
}

