% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_gdp.R
\name{GDP}
\alias{GDP}
\alias{rate_GDP}
\alias{tasa_PIB}
\title{GDP growth rate}
\usage{
rate_GDP(token)

tasa_PIB(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns GDP growth rate vs. same period a year earlier. Wrapper of \code{serie_inegi()} and \code{YoY()}.
}
\note{
BIE route: Indicadores económicos de coyuntura ... Producto interno bruto trimestral, base 2008 ... Series originales ... Valores a precios de 2008 ... Producto interno bruto, a precios de mercado
}
\examples{
\dontrun{
token <- "webservice_token"
# interchangeable
gdp <- rate_GDP(token)
gdp <- tasa_PIB(token)
}
}
\author{
Eduardo Flores
}

