% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_sakbe.R
\name{get_gas}
\alias{get_gas}
\title{Gets gas cost from INEGI API}
\usage{
get_gas(token, onlyPremium = FALSE)
}
\arguments{
\item{token}{Sakbe API token supplied by INEGI}

\item{onlyPremium}{Only export premium price}
}
\value{
data.frame
}
\description{
Helper function
}
