% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAuxiliares_denue.R
\name{hacer_grid}
\alias{hacer_grid}
\title{Proporciona un set de coordenadas}
\usage{
hacer_grid(lat1, lat2, lon1, lon2, espacio_lat = 0.07, espacio_lon = 0.07)
}
\arguments{
\item{lat1}{Esquina 1 de cuadro o área en latitud.}

\item{lat2}{Esquina 2 de cuadro o área en latitud.}

\item{lon1}{Esquina 1 de cuadro o área en longitud.}

\item{lon2}{Esquina 2 de cuadro o área en longitud.}

\item{espacio_lat}{Espacio entre coordenadas, en latitud, por default = 0.07 grados.}

\item{espacio_lon}{Espacio entre coordenadas, en longitud, por default = 0.07 grados.}
}
\value{
Data.frame
}
\description{
Regresa un set de coordenadas que juntas se sobrelapan para crear un cuadro mayor a 5 kilómetros. Posible gracias a Arturo Cardenas \url{https://github.com/arturocm}.
}
\examples{
\dontrun{
token<-"webservice_token"
latitud1 <- "25.669194"
latitud2 <- "25.169194"
longitud1 <- "-100.30990"
longitud2 <- "-101.20102"
varias_coordenadas <- hacer_grid(latitud1, latitud2, longitud1, longitud2)
}

}
\author{
Arturo Cardenas
}
\seealso{
denue_grid
}

