% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_IGAE.R
\name{IGAE}
\alias{IGAE}
\alias{rate_IGAE}
\alias{tasa_IGAE}
\title{IGAE growth rate}
\usage{
rate_IGAE(token)

tasa_IGAE(token)
}
\arguments{
\item{token}{token API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returnes IGAE (Indicador Global de Actividad Economica) growth rate year over year and month over month.
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\details{
Month over month is a seasonally-adjusted series, while the original is used for year over year.
}
\examples{
\dontrun{
token<-"webservice_token"
igae <- rate_IGAE(token)
}
}
\author{
Eduardo Flores
}
