% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_aux.R
\name{incat_neumonics}
\alias{incat_neumonics}
\title{Obtains the catalog of economic data neumonics}
\usage{
incat_neumonics()
}
\value{
data.frame
}
\description{
Neumonics are shorthand names for series of economic data, akin to the Fed FRED names. This catalog downloads the neumonic with the indicator code, to easily match with an API call. This catalog is maintained by INEGI.
}
\examples{
\dontrun{
catalog <- incat_neumonics()
}
}
\author{
Eduardo Flores
}
