% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_bop.R
\name{inegi_bop}
\alias{inegi_bop}
\title{Balance of Payments for Mexico}
\usage{
inegi_bop(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns Current Account revenue, expenses and total and Financial Account total, errors, reservs and adjustments for Mexico.
}
\examples{
\dontrun{
token <- "webservice_token"
balance_of_payments <- inegi_bop(token)
}
}
\author{
Eduardo Flores
}
