% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_aux.R
\name{inegi_code}
\alias{inegi_code}
\title{INEGI code to call}
\usage{
inegi_code(id)
}
\arguments{
\item{id}{numeric INEGI id number}
}
\value{
string
}
\description{
This function is a simple paste0() command that allows you to only pass a unique indicator ID to inegi_series(), instead of the entire URL string.
}
\note{
Works only for national statistics (00000 geography code)
}
\examples{
# Get the corresponding URL for GDP
GPD_ID <- 381016
GDP_CALL_URL <- inegi_code(381016)
}
\author{
Eduardo Flores
}
