% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_routes.R
\name{inegi_destiny}
\alias{inegi_destiny}
\title{Returns INEGI destiny id's with matching searches}
\usage{
inegi_destiny(search, token)
}
\arguments{
\item{search}{Character vector to search for}

\item{token}{Sakbe API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns data.frame with id's and coordinates that match with the API names.
}
\examples{
# All id's in Monterrey, Mex.
\dontrun{
token <- "webservice_token"
dest_ids <- inegi_destiny("monterrey", token)
}

}
\author{
Eduardo Flores
}
