% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{inegi_lines}
\alias{inegi_lines}
\title{Returns INEGI line ID}
\usage{
inegi_lines(latitud, longitud, scale = 1e+05, keep_trying = TRUE, token)
}
\arguments{
\item{latitud}{latitud vector in decimals}

\item{longitud}{longitud vector in decimals}

\item{scale}{scale of map visualization (1:scale)}

\item{keep_trying}{if TRUE the function will start from scale downwards by a degree of 10 until a line is reached}

\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns data.frame with id's that match coordinates.
}
\examples{
# All id's in Monterrey, Mex.
\dontrun{
token <- "webservice_token"
line_ids <- inegi_lines(latitud, longitud, token = token)
}

}
\author{
Eduardo Flores
}

