% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_partner_exports.R
\name{inegi_partner_exports}
\alias{inegi_partner_exports}
\title{Exports of Mexico to trade partners}
\usage{
inegi_partner_exports(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns exports to main trading partners of all products. Regions are the following: United States, Canada, China, CentralAmerica, SouthAmerica
}
\examples{
\dontrun{
token <- "webservice_token"
xbycountry <- inegi_partner_exports(token)
}
}
\author{
Eduardo Flores
}
