% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_sectors.R
\name{inegi_sectors}
\alias{inegi_sectors}
\title{Sectorial activity indices}
\usage{
inegi_sectors(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns indexes of economic sector as defined in INEGI (subsectors of IGAE). 
None of the series are seasonally adjusted.
}
\examples{
\dontrun{
token <- "webservice_token"
sectors <- inegi_sectors(token)
}
}
\author{
Eduardo Flores
}
