% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_student_index.R
\name{inegi_stind}
\alias{inegi_stind}
\title{Student Price Index}
\usage{
inegi_stind(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns the student price index. See \url{http://enelmargen.org/ds/ipe/} for more information.
}
\examples{
\dontrun{
token <- "webservice_token"
studentinflation <- inegi_stind(token)
}
}
\author{
Eduardo Flores
}
