% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_tot.R
\name{inegi_tot}
\alias{inegi_tot}
\title{Terms of trade for Mexico}
\usage{
inegi_tot(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns the terms of trade for Mexico, defined as the price index of exports over the price index of imports.
}
\examples{
\dontrun{
token <- "webservice_token"
tot <- inegi_tot(token)
} 
}
\author{
Eduardo Flores
}
