% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_trade_balance.R
\name{inegi_tradebal}
\alias{inegi_tradebal}
\title{Trade balance}
\usage{
inegi_tradebal(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns exports, imports and trade balance (all products, services and countries) in a data.frame.
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
external_com <- inegi_tradebal(token)
}
}
\author{
Eduardo Flores
}
