\name{inegiR-package}
\alias{inegiR-package}
\alias{inegiR}
\docType{package}
\title{
Integrate INEGI’s (Mexican Stats Office) API with R
}
\description{
Provides functions to download and parse information from INEGI (Official Mexican statistics agency). 
}
\details{
\tabular{ll}{
Package: \tab inegiR\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2018-03-27\cr
}
}
\author{
Eduardo Flores
<eduardo@enelmargen.org>
}
\references{
The INEGI API can be found here: 
http://www.inegi.org.mx/desarrolladores/indicadores/apiindicadores.aspx
}
\keyword{ package }
