% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{inflacion_ciudades}
\alias{inflacion_ciudades}
\title{Obtener inflacion por Ciudad}
\usage{
inflacion_ciudades(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene la tasa de inflación mensual por ciudad.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\note{
Encoding no permite acentos en titulo de descripción
}
\examples{
InflacionCiudades<-inflacion_ciudades(token)
}
\author{
Eduardo Flores
}

