% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{Inflacion_Estudiantes}
\alias{Inflacion_Estudiantes}
\title{Obtener tasa de inflación de Estudiantes}
\usage{
Inflacion_Estudiantes(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API.}
}
\value{
Data.frame con dos columnas: fechas y valores.
}
\description{
Obtiene tasa de inflación de estudiantes, inter anual en porcentaje. Es un wrapper de las funciones Serie_Inegi() y YoY().
La metodología del índice se puede encontrar aquí: http://www.enelmargen.org/2011/04/indice-de-precios-estudiantes.html
}
\examples{
InflacionEstudiantes<-Inflacion_Estudiantes(token)
}
\author{
Eduardo Flores
}

