% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{Inflacion_General}
\alias{Inflacion_General}
\title{Obtener tasa de inflación}
\usage{
Inflacion_General(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API de indicadores.}
}
\value{
Vector numerico
}
\description{
Obtiene tasa de inflación inter anual en porcentaje. Es un wrapper de las funciones Serie_Inegi() y YoY().
}
\examples{
Inflacion<-Inflacion_General(token)
}
\author{
Eduardo Flores
}

