% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{Inflacion_ToT}
\alias{Inflacion_ToT}
\title{Obtener tminos de intercambio}
\usage{
Inflacion_ToT(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame con dos columnas: Fechas y Valores.
}
\description{
Obtiene la razón de términos de intercambio para México (ToT). Es un wrapper de las funciones Serie_Inegi() y YoY().
La razón se define como el índice de precios de exportaciones entre el índice de precios de importaciones.
}
\examples{
TerminosIntercambio<-Inflacion_ToT(token)
}
\author{
Eduardo Flores
}

