% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_inflation.R
\name{inflation}
\alias{inflacion_general}
\alias{inflation}
\alias{overall_inflation}
\title{Get rate of inflation}
\usage{
overall_inflation(token)

inflacion_general(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns anual inflation rate (national, overall rate). Technically, it is the percent anual change of the INPC index.
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
inflation <- overall_inflation(token)
}
}
\author{
Eduardo Flores
}

