% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_commonindicators.R
\name{inind_}
\alias{inind_}
\alias{inind_commerce}
\alias{inind_auto}
\alias{inind_gdp}
\alias{inind_fx}
\alias{inind_unemp}
\alias{inind_prices}
\title{INEGI Direct Indicators}
\usage{
inind_commerce(token)

inind_auto(token)

inind_gdp(token)

inind_fx(token)

inind_unemp(token)

inind_prices(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns common indicators, for simplicity. Will return as a list, with metadata and tibble time dataframe.
}
\details{
inind_commerce = terciary industrial activity (commercial activity monthly). 
inind_auto = auto production. 
innind_gpd = Gross Domestic Product.
inind_fx = USDMXN Exchange rate.
inind_unemp = Unemployment rate.
inind_prices = National price index (for inflation).
}
\examples{
\dontrun{
token<-"webservice_token"
commerce_rate <- inind_commerce(token)
}
}
\author{
Eduardo Flores
}
