% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_legacy.R
\name{legacy}
\alias{legacy}
\alias{series_actividad_industrial}
\alias{series_opiniones}
\title{Miscellaneous functions in spanish}
\usage{
series_opiniones(token)

series_actividad_industrial(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Each of these functions return data in the previous (v1.2) format in spanish and have no interchangeable pair in english. 
See the examples for each series meaning.
}
\note{
Legacy functions, will return data in spanish.
}
\examples{
\dontrun{
token<-"webservice_token"
# Opinion of bussinesmen in different sectors
dat <- series_opiniones(token)
# Industrial activity (monthly survey)
dat <- series_actividad_industrial(token)
}
}
\author{
Eduardo Flores
}

