% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_legacy.R
\name{legacy_productivity}
\alias{legacy_productivity}
\alias{series_productividad_const}
\alias{series_productividad_man}
\title{Productivity functions in spanish}
\usage{
series_productividad_man(token)

series_productividad_const(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Each of these functions return data in the previous (v1.2) format in spanish and have no interchangeable pair in english. 
See the examples for each series meaning.
}
\note{
Legacy functions, will return data in spanish.
}
\examples{
\dontrun{
token <- "webservice_token"
# Manufacturing productivity, by state
dat <- series_productividad_man(token)
# Construction productivity, by state
dat <- series_productividad_const(token)
}
}
\author{
Eduardo Flores
}

