% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_make_grid.R
\name{make_grid}
\alias{make_grid}
\alias{make_grid}
\alias{hacer_grid}
\title{Makes a grid set of coordinates}
\usage{
make_grid(lat1, lat2, lon1, lon2, espacio_lat = 0.07, espacio_lon = 0.07)

hacer_grid(lat1, lat2, lon1, lon2, espacio_lat = 0.07, espacio_lon = 0.07)
}
\arguments{
\item{lat1}{First corner (latitud)}

\item{lat2}{Second corner (latitud)}

\item{lon1}{First corner (longitud)}

\item{lon2}{Second corner (longitud)}

\item{espacio_lat}{Space between latitud coordinates defaults to 0.07 degrees}

\item{espacio_lon}{Space between longitud coordinates defaults to 0.07 degrees}
}
\value{
Data.frame
}
\description{
Returns a set of coordinates that intertwine to create an area larger than 5 kilometers. 
Suggestion by Arturo Cardenas \url{https://github.com/arturocm}.
}
\examples{
\dontrun{
token<-"webservice_token"
latitud1 <- "25.669194"
latitud2 <- "25.169194"
longitud1 <- "-100.30990"
longitud2 <- "-101.20102"
# in spanish (legacy)
setofcoords <- hacer_grid(latitud1, latitud2, longitud1, longitud2)
# in english
setofcoords <- make_grid(latitud1, latitud2, longitud1, longitud2)
}
}
\seealso{
denue_grid
}
\author{
Arturo Cardenas
}
