% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Opiniones}
\alias{Opiniones}
\title{Obtener opiniones empresariales por sector}
\usage{
Opiniones(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene principales componentes de encuestas de Opinión Empresarial del INEGI dividido en 3 sectores: Comercio, Manufacturas y Construcción.
Es un wrapper de las funciones \code{Serie_Inegi()} y \code{YoY()}.
}
\examples{
OpinionMexicanos<-Opiniones(token)
}
\author{
Eduardo Flores
}

