% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_aux.R
\name{ordenar_porconteo}
\alias{ordenar_porconteo}
\title{Order factors by count}
\usage{
ordenar_porconteo(df, col)
}
\arguments{
\item{df}{Data.frame}

\item{col}{Columna with factor. (Bare, no parenthesis).}
}
\value{
data.frame
}
\description{
Wrapper for rapid ordering in a data.frame.
}
\examples{
df <- data.frame(factors=c("A","A","B","C","C","D","A","A"),
                 others=c(1,3,2,4,5,1,2,7))
#order by count
ByCount <- ordenar_porconteo(df, factors)

}
\author{
Eduardo Flores
}
\seealso{
denue_varios_stats
}

