% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rate_cconfidence}
\alias{rate_cconfidence}
\title{Consumer confidence}
\usage{
rate_cconfidence(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns the rate of change in consumer confidence. 
Month over month is a seassonally adjusted rate, year over year is the original series.
}
\examples{
\dontrun{
token<-"webservice_token"
conf <- rate_cconfidence(token)
}
}
\author{
Eduardo Flores
}

