% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rate_commerce}
\alias{rate_commerce}
\title{Commerce growth rate}
\usage{
rate_commerce(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns commerce growth rate (terciary activity) vs. same month year earlier. 
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Commerce <- rate_commerce(token)
}
}
\author{
Eduardo Flores
}

