% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rate_GDP}
\alias{rate_GDP}
\title{GDP growth rate}
\usage{
rate_GDP(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns GDP growth rate vs. same period a year earlier. Wrapper of \code{serie_inegi()} and \code{YoY()}.
}
\note{
BIE route: Indicadores económicos de coyuntura ... Producto interno bruto trimestral, base 2008 ... Series originales ... Valores a precios de 2008 ... Producto interno bruto, a precios de mercado
}
\examples{
\dontrun{
token<-"webservice_token"
gdp <- rate_GDP(token)
}
}
\author{
Eduardo Flores
}

