% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{rate_unemployment}
\alias{rate_unemployment}
\title{Urban unemployment rate}
\usage{
rate_unemployment(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns a unified data series (32 city aggregate) of urban unemployment rate.
Wrapper of \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Desempleo <- rate_unemployment(token)
}
}
\author{
Eduardo Flores
}

