% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_balanza_comercial}
\alias{series_balanza_comercial}
\title{Obtener balanza comercial}
\usage{
series_balanza_comercial(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene exportaciones, importaciones y balance de los dos en un mismo data.frame por mes.
Todos los productos y todos los países. 
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
ComercioExterior<-series_balanza_comercial(token)
}
}
\author{
Eduardo Flores
}

