% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_balanza_pagos}
\alias{series_balanza_pagos}
\title{Obtener Balanza de Pagos}
\usage{
series_balanza_pagos(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene principales componentes de la Balanza de Pagos: 2 de la Cuenta Corriente, 3 de la Cuenta Financiera y sus 2 resultados.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
BalanzadePagosMexico<-series_balanza_pagos(token)
} 
}
\author{
Eduardo Flores
}

