% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_exportaciones_pais}
\alias{series_exportaciones_pais}
\title{Obtener exportaciones por paises}
\usage{
series_exportaciones_pais(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene exportaciones de principales socios comerciales.
Todos los productos y Estados Unidos, Canadá, China, CentroAmerica y América del Sur. 
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
ExportacionesMx<-series_exportaciones_pais(token)
}
}
\author{
Eduardo Flores
}

