% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_ITAE_estados}
\alias{series_ITAE_estados}
\title{Obtiene ITAE}
\usage{
series_ITAE_estados(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene series originales de Indicador Trimestral de Actividad Económica (ITAE) por estado.
}
\examples{
\dontrun{
token<-"webservice_token"
ITAE<-series_ITAE_estados(token)
}
}
\author{
Eduardo Flores
}
\seealso{
series_PIB_estados, series_crecimiento_regiones
}

