% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_PIB_estados}
\alias{series_PIB_estados}
\title{Obtener PIB por Entidad Federativa}
\usage{
series_PIB_estados(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene series originales del PIB a Precios 2008 por Entidad Federativa. Unidades: millones de pesos a precios de 2008.
Es un wrapper de \code{serie_inegi()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Estados<-series_PIB_estados(token)
}
}
\author{
Eduardo Flores
}
\seealso{
series_crecimiento_regiones
}

