% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_produccion_autos}
\alias{series_produccion_autos}
\title{Obtener Produccion de Autos}
\usage{
series_produccion_autos(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene producción automotriz en México y cambio porcentual anual.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
ProduccionAutos<-series_produccion_autos(token)
}
}
\author{
Eduardo Flores
}

