% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_productividad_const}
\alias{series_productividad_const}
\title{Obtener productividad de construcci贸n por estado}
\usage{
series_productividad_const(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene la productivdad de la construcci贸n generada por estado. Se calcula a partir de las series de personal ocupado y de valor de la producci贸n de 
la Encuesta Nacional de Empresas Constructoras (ENEC) al dividirse valor (en miles de pesos a precios corrientes) entre personal (personas).
Es un wrapper de la funci贸n \code{serie_inegi()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Productividad <- series_productividad_const(token)
}
}
\author{
Eduardo Flores
}

