% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_productividad_man}
\alias{series_productividad_man}
\title{Obtener productividad de manufactura por estado}
\usage{
series_productividad_man(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene la productivdad de la manufactura por estado. Se calcula a partir de las series de personal ocupado y de valor de la producción de 
la Encuesta Mensual de la Industria Manufacturera (EMIM) al dividirse valor (en miles de pesos) entre personal (personas).
Es un wrapper de la función \code{serie_inegi()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Productividad <- series_productividad_man(token)
}
}
\author{
Eduardo Flores
}

