% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{series_tipocambio}
\alias{series_tipocambio}
\title{Obtener Tipo de Cambio Peso - USD}
\usage{
series_tipocambio(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene tipo de cambio interbancario (venta) histórico de pesos a dólares.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
USD<-series_tipocambio(token)
}
}
\author{
Eduardo Flores
}

