% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Set_Token.R
\name{Set_Token}
\alias{Set_Token}
\title{Actualiza Token en todas las funciones y metódos}
\usage{
Set_Token(token)
}
\arguments{
\item{token}{Vector de clase caracter con token de INEGI. Ejemplo: "a6hB8912mxn8AJ"}
}
\description{
Al correr la función, el token que se elige se actualiza en todos los catalogos y funciones que lo usan.
Esta función sirve para evitar que se detenga un proceso por un token equivocado.
}
\examples{
Set_Token("tokenProporcionadoporWebservice")
}

