% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_student_inflation.R
\name{student_inflation}
\alias{inflacion_estudiantes}
\alias{student_inflation}
\title{Student Price Index Inflation}
\usage{
inflacion_estudiantes(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
Data.frame
}
\description{
Returns the calculated inflation for students. See \url{http://enelmargen.org/eem/ipe/} for more information.
}
\examples{
\dontrun{
token<-"webservice_token"
studentinflation <- inflacion_estudiantes(token)
}
}
\author{
Eduardo Flores
}

