% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tasas.R
\name{Tasa_Comercio}
\alias{Tasa_Comercio}
\title{Obtener Tasa de Crecimiento de Comercio}
\usage{
Tasa_Comercio(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene tasa de crecimiento del Comercio (Actividad Terciaria), por mes.
Es un wrapper de las funciones \code{Serie_Inegi()} y \code{YoY()}.
}
\examples{
Comercio<-Tasa_Comercio(token)
}
\author{
Eduardo Flores
}

