% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Tasa_Confianza}
\alias{Tasa_Confianza}
\title{Obtener Confianza del Consumidor}
\usage{
Tasa_Confianza(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Cambio de Confianza del Consumidor
Devuelve tasas de serie desestacionalizada anual, desestacionalizada contra mes previo y serie original anual.
Es un wrapper de las funciones \code{Serie_Inegi()} y \code{YoY()}.
}
\examples{
ConfianzaEconomia<-Tasa_Confianza(token)
}
\author{
Eduardo Flores
}

