% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tasas.R
\name{Tasa_IGAE}
\alias{Tasa_IGAE}
\title{Obtener IGAE}
\usage{
Tasa_IGAE(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Crecimiento de Indicador Global de Actividad Económica
Devuelve tasas de serie desestacionalizada anual, desestacionalizada contra mes previo y serie original anual.
}
\examples{
ActividadEconomica<-Tasa_IGAE(token)
}
\author{
Eduardo Flores
}

