% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tasas.R
\name{Tasa_SectoresYoY}
\alias{Tasa_SectoresYoY}
\title{Obtener cambios porcentuales por sector}
\usage{
Tasa_SectoresYoY(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Crecimiento de Indicador Global de Actividad Económica por subsector.
Todas las tasas son con series originales. Cambio porcentual anual.
}
\examples{
Sectores<-Tasa_SectoresYoY(token)
}
\author{
Eduardo Flores
}

