% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TasaPIB_Estados_Inegi.R
\name{TasaPIB_Estados_Inegi}
\alias{TasaPIB_Estados_Inegi}
\title{Obtiene las tasas de crecimiento del PIB por Estado}
\usage{
TasaPIB_Estados_Inegi(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API.}
}
\value{
Data.frame numerico con 33 columnas (1 por estado y fecha)
}
\description{
Obtiene la tasa de crecimiento vs. mismo periodo de un año antes en porcentaje de cada estado del país (en columna) Es un wrapper de las funciones Serie_Inegi() y YoY().
El orden de los estados es orden alfabetico.
}
\examples{
PIBEstados <- TasaPIB_Estados_Inegi(token)
}
\author{
Eduardo Flores
}

