% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TasaPIB_Inegi.R
\name{TasaPIB_Inegi}
\alias{TasaPIB_Inegi}
\title{Obtener tasa de crecimiento del PIB}
\usage{
TasaPIB_Inegi(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API.}
}
\value{
Vector numerico
}
\description{
Obtiene tasa de crecimiento vs. mismo periodo de un año antes en porcentaje. Es un wrapper de las funciones Serie_Inegi() y YoY().
}
\note{
Ruta tematica BIE: Indicadores económicos de coyuntura ... Producto interno bruto trimestral, base 2008 ... Series originales ... Valores a precios de 2008 ... Producto interno bruto, a precios de mercado
}
\examples{
CrecimientoMex<-TasaPIB_Inegi(token)
}
\author{
Eduardo Flores
}

