% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_trade_balance.R
\name{trade_balance}
\alias{trade_balance}
\alias{trade_balance}
\alias{series_balanza_comercial}
\title{Trade balance}
\usage{
trade_balance(token)

series_balanza_comercial(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns exports, imports and trade balance (all products, services and countries) in a data.frame.
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
external_com <- trade_balance(token)
}
}
\author{
Eduardo Flores
}
