% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_aux.R
\name{ultimos}
\alias{ultimos}
\title{Returns n most recient data points}
\usage{
ultimos(serie, col = "Fechas", n = 12)
}
\arguments{
\item{serie}{serie in data.frame}

\item{col}{Column with dates}

\item{n}{amount of periods}
}
\value{
data.frame
}
\description{
Wrapper for other functions
}
\examples{
#return last 13 months
\dontrun{
Ultimos <- ultimos(Inflation, n = 12)
}

}
\author{
Eduardo Flores
}
\seealso{
denue_varios_stats
}

