% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuncionesAuxiliares.R
\name{UnirFechas}
\alias{UnirFechas}
\title{Unir series por fecha}
\usage{
UnirFechas(lista)
}
\arguments{
\item{serie}{vector o serie de tiempo con datos númericos}

\item{lapso}{separaciones por año a contemplar (12 = datos mensuales, 4 = datos trimestrales)}

\item{decimal}{¿Quieres que el resultado este en decimales? Default = TRUE. False obtiene el decimal x 100.}
}
\value{
Vector numerico
}
\description{
Permite unir tres o más series de indicadores en un data.frame con fechas. Requiere de una lista y diferentes nombres por indicador.
}
\note{
Las series deben de tener diferentes nombres, excepto columna a unir (Fechas). El argumento es de tipo Lista. El argumento debe ser explícitamente anunciad
}
\examples{
#Unir Crecimiento, Inflación y Balanza comercial
token<-"abc123"
Inflacion<-Inflacion_General(token)
Balanza<-Balanza_Comercial(token)
Actividad<-Tasa_IGAE(token)
ReporteUnido<-UnirFechas(lista=list(Inflacion,Balanza,Actividad))
}
\author{
Eduardo Flores
}

