% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Autos}
\alias{Autos}
\title{Obtener Producción de Autos}
\usage{
Autos(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame con 3 columnas
}
\description{
Obtiene producción automotriz en México y cambio porcentual anual.
}
\examples{
ProduccionAutos<-Autos(token)
}
\author{
Eduardo Flores
}

