% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_balance_payments.R
\name{balance_payments}
\alias{balance_payments}
\alias{balance_payments}
\alias{series_balanza_pagos}
\title{Balance of Payments}
\usage{
balance_payments(token)

series_balanza_pagos(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns the main components of balance of payments.
Wrapper for \code{inegi_series()} and \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
MxBoP <- balance_payments(token)
} 
}
\author{
Eduardo Flores
}
