% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Balanza_Comercial}
\alias{Balanza_Comercial}
\title{Obtener balanza comercial}
\usage{
Balanza_Comercial(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API de indicadores.}
}
\value{
data.frame con 4 columnas
}
\description{
Obtiene exportaciones, importaciones y balance de los dos en un mismo data.frame por mes.
Todos los productos y todos los países
}
\examples{
ComercioExterior<-Balanza_Comercial(token)
}
\author{
Eduardo Flores
}

