% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuncionesAuxiliares.R
\name{crecer}
\alias{crecer}
\title{Crece una serie por tasas}
\usage{
crecer(tasas, comienzo)
}
\arguments{
\item{tasas}{vector con tasas de crecimiento}

\item{comienzo}{número inicial}
}
\value{
Vector numerico
}
\description{
Al especificar un dato inicial, "crece" una serie de datos usando un vector de tasas de crecimiento. La tasa se hace de periodo en periodo.
}
\examples{
tasas_crecimiento<-c(1.10,1.20,1.05,1.02,1.10)

# Crecer por esas tasas (en cada periodo) el 100:
Resultados<-crecer(tasas = tasas_crecimiento, comienzo = 100)
}
\author{
Eduardo Flores
}
\seealso{
series_crecimiento_regiones
}

