% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{Exportaciones_Pais}
\alias{Exportaciones_Pais}
\title{Obtener exportaciones por principal país}
\usage{
Exportaciones_Pais(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API de indicadores.}
}
\value{
data.frame con 6 columnas
}
\description{
Obtiene exportaciones de principales socios comerciales.
Todos los productos y Estados Unidos, Canadá, China, CentroAmerica y América del Sur.
}
\examples{
ExportacionesMx<-Exportaciones_Pais(token)
}
\author{
Eduardo Flores
}

