% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OtrosIndicadores.R
\name{exportaciones_pais}
\alias{exportaciones_pais}
\title{Obtener exportaciones por paises}
\usage{
exportaciones_pais(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene exportaciones de principales socios comerciales.
Todos los productos y Estados Unidos, Canadá, China, CentroAmerica y América del Sur.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\note{
Encoding no permite acéntos en título de descripción
}
\examples{
ExportacionesMx<-exportaciones_pais(token)
}
\author{
Eduardo Flores
}

