\name{inegiR-package}
\alias{inegiR-package}
\alias{inegiR}
\docType{package}
\title{
Funciones para obtener datos de INEGI
}
\description{
Funciones para obtener, interactuar y gráficar datos del Instituto Nacional de Estadística y Geografía (México) via el servicio web de SDMX. La implementación se basa en wrappers de xml, rcurl y rsdmx. 
}
\details{
\tabular{ll}{
Package: \tab inegiR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-19\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Eduardo Flores
<eduardo@enelmargen.org>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
