% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{inflacion_estudiantes}
\alias{inflacion_estudiantes}
\title{Obtener tasa de inflacion de Estudiantes}
\usage{
inflacion_estudiantes(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene tasa de inflación de estudiantes, inter anual en porcentaje. Es un wrapper de las funciones Serie_Inegi() y YoY().
La metodología del índice se puede encontrar aquí: \url{http://enelmargen.org/eem/ipe/}
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
InflacionEstudiantes<-inflacion_estudiantes(token)
}
}
\author{
Eduardo Flores
}

