% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{inflacion_general}
\alias{inflacion_general}
\title{Obtener tasa de inflacion}
\usage{
inflacion_general(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API de indicadores.}
}
\value{
Data.frame
}
\description{
Obtiene tasa de inflación inter anual en porcentaje.
La inflación se define como el cambio porcentual en el INPC. 
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Inflacion<-inflacion_general(token)
}
}
\author{
Eduardo Flores
}

