% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndicesPrecios.R
\name{inflacion_tot}
\alias{inflacion_tot}
\title{Obtener terminos de intercambio}
\usage{
inflacion_tot(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene la razón de términos de intercambio para México (ToT). Es un wrapper de las funciones serie_inegi() y YoY(). 
La razón se define como el índice de precios de exportaciones entre el índice de precios de importaciones. 
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
TerminosIntercambio<-inflacion_tot(token)
} 
}
\author{
Eduardo Flores
}

