% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuncionesAuxiliares.R
\name{ordenar_porconteo}
\alias{ordenar_porconteo}
\title{Ordenar por conteo de factores}
\usage{
ordenar_porconteo(df, col)
}
\arguments{
\item{df}{Data.frame a condensar}

\item{col}{Columna con factores. Se pone sin parentesis.}
}
\value{
Data.frame
}
\description{
Wrapper para ordenar rapidamente de mayor a menor por grupos un data.frame.
}
\examples{
df<-data.frame(factores=c("A","A","B","C","C","D","A","A"),otros=c(1,3,2,4,5,1,2,7))

#Ordenar, de mayor a menor, por conteo de factores
PorConteo<-ordenar_porconteo(df, factores)
}
\author{
Eduardo Flores
}
\seealso{
denue_varios_stats
}

