% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Serie_Inegi.R
\name{Serie_Inegi}
\alias{Serie_Inegi}
\title{Obtiene serie de tiempo de INEGI}
\usage{
Serie_Inegi(serie, token, metadata = FALSE)
}
\arguments{
\item{serie}{Vector en caracter de url de dirección. Este es un metódo directo (se requiere de URL en formato XML, con token)}

\item{metadata}{Default = FALSE, si TRUE, trae columnas con Región, Unidad, Indicador (# INEGI) y Frecuencia.}
}
\value{
Dataframe
}
\description{
Regresa Data.Frame con la serie de tiempo escogida, al buscar en el webservice del INEGI y parsear via RSDMX y ZOO.
Si parametro Metadata=TRUE, regresa además Región, Unidad, Indicador (# INEGI) y Frecuencia.
Es una de las funciones primitivas del paquete.
}
\examples{
#Serie de INPC General
token<-"tokenProporcionadoporWebservice"
url <- "http://www3.inegi.org.mx/sistemas/api/indicadores/v1//Indicador/216064/00000/es/false/xml/"
Serie <- Serie_Inegi(url,token)
}
\author{
Eduardo Flores
}

