% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasas.R
\name{tasa_confianza}
\alias{tasa_confianza}
\title{Obtener Confianza del Consumidor}
\usage{
tasa_confianza(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Cambio de Confianza del Consumidor
Devuelve tasas de serie desestacionalizada anual, desestacionalizada contra mes previo y serie original anual.
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
ConfianzaEconomia<-tasa_confianza(token)
}
}
\author{
Eduardo Flores
}

