% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tasas.R
\name{Tasa_Desempleo}
\alias{Tasa_Desempleo}
\title{Obtener Desempleo Urbano}
\usage{
Tasa_Desempleo(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame con 2 columnas
}
\description{
Obtiene tasa de desocupación (serie unificada) urbana (agregado de 32 ciudades)
}
\examples{
Desempleo<-Tasa_Desempleo(token)
}
\author{
Eduardo Flores
}

