% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Tasas.R
\name{tasa_desempleo}
\alias{tasa_desempleo}
\title{Obtener Desempleo Urbano}
\usage{
tasa_desempleo(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene tasa de desocupación (serie unificada) urbana (agregado de 32 ciudades)
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\examples{
\dontrun{
token<-"webservice_token"
Desempleo<-tasa_desempleo(token)
}
}
\author{
Eduardo Flores
}

