% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasas.R
\name{tasa_IGAE}
\alias{tasa_IGAE}
\title{Obtener IGAE}
\usage{
tasa_IGAE(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Crecimiento de Indicador Global de Actividad Económica
Devuelve tasas de serie desestacionalizada anual, desestacionalizada contra mes previo y serie original anual.
}
\examples{
\dontrun{
token<-"webservice_token"
ActividadEconomica<-tasa_IGAE(token)
}
}
\author{
Eduardo Flores
}

