% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasas.R
\name{tasa_PIB}
\alias{tasa_PIB}
\title{Obtener tasa de crecimiento del PIB}
\usage{
tasa_PIB(token)
}
\arguments{
\item{token}{token persona emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene tasa de crecimiento vs. mismo periodo de un año antes en porcentaje. 
Es un wrapper de las funciones \code{serie_inegi()} y \code{YoY()}.
}
\note{
Ruta tematica BIE: Indicadores económicos de coyuntura ... Producto interno bruto trimestral, base 2008 ... Series originales ... Valores a precios de 2008 ... Producto interno bruto, a precios de mercado
}
\examples{
\dontrun{
token<-"webservice_token"
CrecimientoMex<-tasa_PIB(token)
}
}
\author{
Eduardo Flores
}

