% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tasas.R
\name{tasa_sectoresYoY}
\alias{tasa_sectoresYoY}
\title{Obtener cambios porcentuales por sector}
\usage{
tasa_sectoresYoY(token)
}
\arguments{
\item{token}{token personal emitido por el INEGI para acceder al API.}
}
\value{
Data.frame
}
\description{
Obtiene Tasas de Crecimiento de Indicador Global de Actividad Económica por subsector. 
Todas las tasas son con series originales. Cambio porcentual anual.
}
\examples{
\dontrun{
token<-"webservice_token"
Sectores<-tasa_sectoresYoY(token)
}
}
\author{
Eduardo Flores
}

