% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_tot.R
\name{terms_trade}
\alias{inflacion_tot}
\alias{terms_trade}
\alias{trade_terms}
\title{Terms of trade}
\usage{
trade_terms(token)

inflacion_tot(token)
}
\arguments{
\item{token}{API token supplied by INEGI}
}
\value{
data.frame
}
\description{
Returns the terms of trade for Mexico, defined as the price index of exports over the price index of imports.
Wrapper for \code{serie_inegi()} and \code{YoY()}.
}
\examples{
\dontrun{
token <- "webservice_token"
tot <- trade_terms(token)
} 
}
\author{
Eduardo Flores
}

