% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAuxiliares.R
\name{YoY}
\alias{YoY}
\title{Calcular tasas de crecimiento}
\usage{
YoY(serie, lapso, decimal = TRUE)
}
\arguments{
\item{serie}{vector o serie de tiempo con datos númericos}

\item{lapso}{separaciones por año a contemplar (12 = datos mensuales, 4 = datos trimestrales)}

\item{decimal}{¿Quieres que el resultado este en decimales? Default = TRUE. False obtiene el decimal x 100.}
}
\value{
Vector numerico
}
\description{
Calcula tasas de crecimiento de una serie.
}
\note{
La serie debe estar en orden asciendiente (Posición inicial es la más antigua). La función de Serie_Inegi() guarda en ese orden.
}
\examples{
#Calcular la inflación (Ver Inflacion_Inegi() para un método más directo)
\dontrun{
token<-"webservice_token"
INPC<-serie_inegi(INPC, token)
Inflacion<-YoY(INPC$Valores,12)
}
}
\author{
Eduardo Flores
}

